decoupling query and expression rewriter
